+++ build.proj
@@ -7,6 +7,9 @@
   <Import Project="build/Microsoft.DotNet.Cli.Monikers.props" />
 
   <PropertyGroup>
+    <Architecture Condition=" '$(OverrideArchitecture)' != '' ">$(OverrideArchitecture)</Architecture>
+    <Rid Condition=" '$(OverrideRid)' != '' ">$(OverrideRid)</Rid>    
+
     <!--
       $(OS) is set to Unix/Windows_NT. This comes from an environment variable on Windows and MSBuild on Unix.
     -->
